moved methods to migration service
